+++ src/server/globalErrorCacheServer.ts
@@ -1,9 +1,10 @@
 /**
  * A global view on the errors. This consolidates errors from all the workers + web server.
  *
- * Note: Only a few error sources at the moment
- * - worker: ts worker
+ * A few error sources at the moment
  * - web server: active project config does some validation
+ * - ts worker
+ * - linter worker
  */
 import {ErrorsCache} from "./utils/errorsCache";
 export const errorsCache = new ErrorsCache();
