+++ src/server/workers/lint/lintMaster.ts
@@ -4,7 +4,7 @@ import * as contract from "./lintContract";
 import {resolve} from "../../../common/utils";
 
 /** This is were we push the errors */
-import {errorsCache} from "../../globalErrorCache";
+import {errorsCache} from "../../globalErrorCacheServer";
 /** This is where we get the active project information from */
 import * as activeProjectConfig from "../../disk/activeProjectConfig";
 import * as projectDataLoader from "../../disk/projectDataLoader";
